[#7] refactor : context api를 적용
- todo 전역 상태 관리
- 적절한 type 적용
