+++ src/App.tsx
@@ -3,30 +3,12 @@ import { layouts as S } from 'styles/layouts';
 import StateArea from 'components/StateArea';
 import Header from 'components/Header';
 import { tagStates } from 'utils/constants';
-import { useTodo } from 'utils/todoService';
-
-import {TodosContextProvider, useTodosState} from 'contexts/Todo/Store'
-import TodoList from 'components/TodoList';
-import TodoForm from 'components/TodoForm';
+import { TodosContextProvider } from 'contexts/Todo/TodoStore';
 
 const App = () => {
-  const {
-    todoState,
-    nextIdState,
-    increamentNextId,
-    removeTodo,
-    createTodo,
-    updateToDo,
-  } = useTodo();
   const [userName, setUserName] = useState('');
   return (
-    <>
     <TodosContextProvider>
-      {/* child */}
-      <TodoForm/>
-      <TodoList/>
-    </TodosContextProvider>
-    
       <S.Wrap>
         <S.Header>
           <Header userName={userName} setUserName={setUserName} />
@@ -34,21 +16,12 @@ const App = () => {
         <S.Main>
           {tagStates.map((v) => (
             <S.Section key={v}>
-              <StateArea
-                tagName={v}
-                userName={userName}
-                nextId={nextIdState}
-                createTodo={createTodo}
-                increamentNextId={increamentNextId}
-                removeTodo={removeTodo}
-                todoState={todoState}
-                updateToDo={updateToDo}
-              />
+              <StateArea tagName={v} userName={userName} />
             </S.Section>
           ))}
         </S.Main>
       </S.Wrap>
-      </>
+    </TodosContextProvider>
   );
 };
