+++ src/components/common/ToDoCreate.tsx
@@ -1,17 +1,24 @@
 import React from 'react';
 import { Modal } from 'components/modal';
 import TaskForm from './TaskForm';
+import { Todo } from 'contexts/Todo/TodoStore';
 
 interface TodoCreateProps {
   isCreate: boolean;
   open: boolean;
   stars: boolean[];
-  setStars: any;
-  setStarIndex: (starIndex: number) => void;
-  handleSave: any;
-  handleCancel: any;
+  setStars: React.Dispatch<React.SetStateAction<boolean[]>>;
+  setStarIndex: React.Dispatch<React.SetStateAction<number>>;
+  handleSave: (
+    e:
+      | React.FormEvent<HTMLFormElement>
+      | React.MouseEvent<HTMLButtonElement, MouseEvent>,
+    isCreate: boolean,
+    todo: Todo,
+  ) => void;
+  handleCancel: (isCreate: boolean) => void;
   inputTask: string;
-  setInputTask: any;
+  setInputTask: React.Dispatch<React.SetStateAction<string>>;
 }
 
 const ToDoCreate = (props: TodoCreateProps) => {
