+++ src/utils/localStorage.ts
@@ -1,4 +1,4 @@
-import { Itodo } from './todoService';
+import { Todo } from 'contexts/Todo/TodoStore';
 
 const getLocalStorage = (key: string) => {
   const data = localStorage.getItem(key);
@@ -35,7 +35,7 @@ export const saveTodoStorage = (todo: any) => {
 
 export const removeTodoStorage = (id: number) => {
   const todos = getTodoStorage();
-  const newTodos = todos.filter((todo: Itodo) => todo.id !== id);
+  const newTodos = todos.filter((todo: Todo) => todo.id !== id);
   saveTodoStorage(newTodos);
   return newTodos;
 };
